ã일기장 상세보기 오ë수정

+++ Ai-Diary-server/src/main/java/com/domain/demo_backend/mapper/UserMapper.java
@@ -32,6 +32,6 @@ public interface UserMapper {
     void updateUser(User existingUser);
 
     void insertKakaoUser(User user);
-    BigInteger findIndexByEmail(String email);
+    Long findIndexByEmail(String email);
 
 }
